Create condition extension to evaluate conditions
